requirements.txt
pyyaml
python-dateutil
MySQL-python
requests
pandas